Fix multi-page display on bootstrap_tabs
If this is a multi-page form i.e. one of the Groups has split_page set,
then groups are displayed consolidated onto tabs until the split_page
group is reached.

bootstrap_tabs currently displays tabs for all groups, but tabs for
non-split_page groups  after first are empty.

This fix puts the value of split_page into a field and only shows the
tab if needed.