Provide PLG_LIST_ENVELOPE_ENVELOPE string
For dropdown menu in list, this plugin asks for
PLG_LIST_ENVELOPE_ENVELOPE which does not exist.

This commit provides this as the same string as PLG_LIST_EMAIL_EMAIL
which seems to be in error (but which is retained in case it is used by
anyone in eval code).