Increase device compatibility
According to the W3C spec, capture is a boolean attribute - if specified
it is on.

However according to
https://dev.opera.com/articles/media-capture-in-mobile-browsers/ some
browsers conform to an older spec.

This commit implements the older spec which should be forward compatible
with the new spec (because capture="camera" is true-ish when capture is
considered a Boolean attribute).